almost step7

+++ passe3.rb
@@ -9,12 +9,16 @@ ActiveRecord::Base.establish_connection(adapter: 'sqlite3', database: 'data.sqli
 class Formation < ActiveRecord::Base
 end
 
-
-
+count = Formation.all.count
+cpt = 0
 Formation.all.each { |formation|
+  outfile = "fiches/#{formation.dest}/#{formation.id}.html"
+  cpt += 1
+  puts "#{cpt}/#{count}"
+  next if File.exist? outfile
   puts "get #{formation.url}"
   doc = Nokogiri::HTML(open(URI.encode(formation.url)))
-  outfile = "fiches/#{formation.dest}/#{formation.id}.html"
   File.write(outfile, doc.inner_html)
-  break
+
+  # break if cpt > 10
 }
